+++ Classes/Domain/Aggregation/JavaScriptErrorDetailsAggregator.php
@@ -26,7 +26,7 @@ final class JavaScriptErrorDetailsAggregator
     }
 
     /**
-     * @param list<array{browserName: string, browserIcon: string, actionDetails: list<array{type: string, eventCategory: string, eventAction: string, url: string, timestamp: int}>}> $details
+     * @param list<array{browserName: string, browserIcon: string, browserVersion: string, actionDetails: list<array{type: string, eventCategory: string, eventAction: string, url: string, timestamp: int}>}> $details
      */
     public function aggregate(array $details): JavaScriptErrorDetails
     {
@@ -39,7 +39,7 @@ final class JavaScriptErrorDetailsAggregator
                     continue;
                 }
 
-                $this->javaScriptDetails->incrementBrowserCount($detail['browserName'], $detail['browserIcon']);
+                $this->javaScriptDetails->incrementBrowserCount($detail['browserName'], $detail['browserIcon'], $detail['browserVersion']);
                 $this->javaScriptDetails->incrementUrlCount($actionDetail['url']);
                 $this->javaScriptDetails->incrementScriptCount($actionDetail['eventAction']);
                 $this->javaScriptDetails->compareAndStoreLastAppearance($actionDetail['timestamp']);