+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Browser version to JavaScript error details modal (#34)
+
 ## [1.5.0] - 2022-06-13
 
 ### Removed